+++ CDN/js/index.js
@@ -1,0 +1,79 @@
+(function(){
+
+	console.log('checkSystemRequirements');
+	console.log(JSON.stringify(ZoomMtg.checkSystemRequirements()));
+
+    //ZoomMtg.setZoomJSLib('https://dmogdx0jrul3u.cloudfront.net/1.3.7/lib', '/av'); option
+
+    ZoomMtg.preLoadWasm();
+
+    ZoomMtg.prepareJssdk();
+
+    var API_KEY = '';
+
+    /**
+     * NEVER PUT YOUR ACTUAL API SECRET IN CLIENT SIDE CODE, THIS IS JUST FOR QUICK PROTOTYPING
+     * The below generateSignature should be done server side as not to expose your api secret in public
+     * You can find an eaxmple in PHP here: https://gist.github.com/joshuawoodward/7574df3df9a089e2663582a2cf9f188b
+     */
+    var API_SECRET = '';
+
+    document.getElementById('join_meeting').addEventListener('click', function(e){
+
+        e.preventDefault();
+
+        if(!this.form.checkValidity()){
+            alert("Enter Name and Meeting Number");
+            return false;
+        }
+
+        var meetConfig = {
+            apiKey: API_KEY,
+            apiSecret: API_SECRET,
+            meetingNumber: parseInt(document.getElementById('meeting_number').value),
+            userName: document.getElementById('display_name').value,
+            passwd: "",
+            leaveUrl: "https://zoom.us",
+            role: 0
+        };
+
+
+        var signature = ZoomMtg.generateSignature({
+            meetingNumber: meetConfig.meetingNumber,
+            apiKey: meetConfig.apiKey,
+            apiSecret: meetConfig.apiSecret,
+            role: meetConfig.role,
+            success: function(res){
+                console.log(res.result);
+            }
+        });
+
+        ZoomMtg.init({
+            leaveUrl: 'http://www.zoom.us',
+            isSupportAV: true,
+            success: function () {
+                ZoomMtg.join(
+                    {
+                        meetingNumber: meetConfig.meetingNumber,
+                        userName: meetConfig.userName,
+                        signature: signature,
+                        apiKey: meetConfig.apiKey,
+                        userEmail: 'email@gmail.com',
+                        passWord: "",
+                        success: function(res){
+                            console.log('join meeting success');
+                        },
+                        error: function(res) {
+                            console.log(res);
+                        }
+                    }
+                );
+            },
+            error: function(res) {
+                console.log(res);
+            }
+        });
+
+    });
+
+})();