+++ Local/js/index.js
@@ -1,0 +1,73 @@
+import { ZoomMtg } from 'zoomus-jssdk';
+
+console.log('checkSystemRequirements');
+console.log(JSON.stringify(ZoomMtg.checkSystemRequirements()));
+
+// ZoomMtg.setZoomJSLib('http://localhost:9999/node_modules/zoomus-jssdk/dist/lib', '/av');
+ZoomMtg.preLoadWasm();
+ZoomMtg.prepareJssdk();
+
+
+const API_KEY = '';
+
+/**
+    * NEVER PUT YOUR ACTUAL API SECRET IN CLIENT SIDE CODE, THIS IS JUST FOR QUICK PROTOTYPING
+    * The below generateSignature should be done server side as not to expose your api secret in public
+    * You can find an eaxmple in PHP here: https://gist.github.com/joshuawoodward/7574df3df9a089e2663582a2cf9f188b
+    */
+const API_SECRET = '';
+
+document.getElementById('join_meeting').addEventListener('click', (e) => {
+    e.preventDefault();
+
+    const meetConfig = {
+        apiKey: API_KEY,
+        apiSecret: API_SECRET,
+        meetingNumber: parseInt(document.getElementById('meeting_number').value, 10),
+        userName: document.getElementById('display_name').value,
+        passwd: '',
+        leaveUrl: 'https://zoom.us',
+        role: 0
+    };
+
+    ZoomMtg.generateSignature({
+        meetingNumber: meetConfig.meetingNumber,
+        apiKey: meetConfig.apiKey,
+        apiSecret: meetConfig.apiSecret,
+        role: meetConfig.role,
+        success(res) {
+            console.log('signature', res.result);
+            ZoomMtg.init({
+                leaveUrl: 'http://www.zoom.us',
+                isSupportAV: true,
+                videoDrag: false,
+                videoHeader: false,
+                sharingMode: 'fit',
+                webEndpoint: meetConfig.webEndpoint,
+                rwcEndpoint: meetConfig.rwcEndpoint,
+                success() {
+                    ZoomMtg.join(
+                        {
+                            meetingNumber: meetConfig.meetingNumber,
+                            userName: meetConfig.userName,
+                            signature: res.result,
+                            apiKey: meetConfig.apiKey,
+                            userEmail: 'email@gmail.com',
+                            passWord: '',
+                            success() {
+                                $('#nav-tool').hide();
+                                console.log('join meeting success');
+                            },
+                            error(res) {
+                                console.log(res);
+                            }
+                        }
+                    );
+                },
+                error(res) {
+                    console.log(res);
+                }
+            });
+        }
+    });
+});